+++ app-web/src/main/java/com/sphincs/web/TripController.java
@@ -14,9 +14,11 @@ import org.springframework.web.servlet.ModelAndView;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Pattern;
 
 @Controller
 @RequestMapping("/trips")
@@ -44,72 +46,81 @@ public class TripController {
     }
 
     @RequestMapping("/tripInputForm")
-    public ModelAndView launchDriverInputFormView() {
+    public ModelAndView launchTripInputFormView() {
         return new ModelAndView("tripInputForm", "trip", new Trip());
     }
 
     @RequestMapping("/submitData")
-    public String getInputFormView(@RequestParam("DriverName") String driverName,
+    public ModelAndView getInputFormView(@RequestParam("DriverName") String driverName,
                                    @RequestParam("StartPoint") String startPoint,
                                    @RequestParam("EndPoint") String endPoint,
-                                   @RequestParam("Distance") Double distance,
+                                   @RequestParam("Distance") String distance,
                                    @RequestParam("StartDate") String startDate,
                                    @RequestParam("EndDate") String endDate) throws ParseException {
-        Trip trip = new Trip();
-        trip.setDriver(driverService.getDriverByName(driverName));
-        trip.setStartPoint(startPoint);
-        trip.setEndPoint(endPoint);
-        trip.setDistance(distance);
-        trip.setStartDate(format.parse(startDate));
-        trip.setEndDate(format.parse(endDate));
-        trip.setSumFuel();
-        tripService.addTrip(trip);
-        return "redirect:/trips/tripsList";
+        LOGGER.debug("add new trip");
+        Trip trip = checkTripFields(new Trip(), driverName, startPoint, endPoint, distance, startDate, endDate);
+        if (trip != null) {
+            tripService.addTrip(trip);
+            return getTripsListView();
+        } else {
+            return getErrorPage("Введены некорректные данные. Проверьте правильность ввода");
+        }
     }
 
     @RequestMapping(value = "/tripRemoveForm")
-    public String removeDriverFormView() {
-        return "removeTrip";
+    public ModelAndView removeDriverFormView() {
+        LOGGER.debug("get all available trips");
+        List<Trip> trips = tripService.getAllTrips();
+        return new ModelAndView("removeTrip", "trips", trips);
     }
 
     @RequestMapping(value = "/tripRemove")
-    public String removeDriverView(@RequestParam("Id") Long tripId) {
+    public ModelAndView removeDriverView(@RequestParam("Id") Long tripId) {
         LOGGER.debug("Delete trip with id = " + tripId);
-        tripService.removeTrip(tripId);
-        return "redirect:/trips/tripsList";
+        Trip trip = tripService.getTripById(tripId);
+        if (trip != null) {
+            tripService.removeTrip(tripId);
+            return getTripsListView();
+        } else {
+            return getErrorPage("Рейс с ID = " + tripId + " не существует");
+        }
     }
 
     @RequestMapping(value = "/tripUpdateId")
-    public String updateDriverFormView() {
-        return "updateTrip";
+    public ModelAndView updateDriverFormView() {
+        LOGGER.debug("get all available trips");
+        List<Trip> trips = tripService.getAllTrips();
+        return new ModelAndView("updateTrip", "trips", trips);
     }
 
     @RequestMapping(value = "/tripUpdateForm")
     public ModelAndView getTripByIdView(@RequestParam("Id") Long tripId) {
-        Trip trip = tripService.getTripById(tripId);
         LOGGER.debug("trip.id = " + tripId);
-        return new ModelAndView("updateTripForm", "trip", trip);
+        Trip trip = tripService.getTripById(tripId);
+        if (trip != null) {
+            return new ModelAndView("updateTripForm", "trip", trip);
+        } else {
+            return getErrorPage("Рейс с ID = " + tripId + " не существует");
+        }
     }
 
     @RequestMapping(value = "/tripUpdate")
-    public String updateDriver(@RequestParam("Id") Long tripId,
+    public ModelAndView updateDriver(@RequestParam("Id") Long tripId,
                                @RequestParam("DriverName") String driverName,
                                @RequestParam("StartPoint") String startPoint,
                                @RequestParam("EndPoint") String endPoint,
-                               @RequestParam("Distance") Double distance,
+                               @RequestParam("Distance") String distance,
                                @RequestParam("StartDate") String startDate,
                                @RequestParam("EndDate") String endDate) throws ParseException {
         LOGGER.debug("update trip with id = " + tripId);
         Trip trip = tripService.getTripById(tripId);
-        trip.setDriver(driverService.getDriverByName(driverName));
-        trip.setStartPoint(startPoint);
-        trip.setEndPoint(endPoint);
-        trip.setDistance(distance);
-        trip.setStartDate(format.parse(startDate));
-        trip.setEndDate(format.parse(endDate));
-        trip.setSumFuel();
-        tripService.updateTrip(trip);
-        return "redirect:/trips/tripsList";
+        trip = checkTripFields(trip, driverName, startPoint, endPoint, distance, startDate, endDate);
+        if (trip != null) {
+            tripService.updateTrip(trip);
+            return getTripsListView();
+        } else {
+            return getErrorPage("Введены некорректные данные. Проверьте правильность ввода");
+        }
     }
 
     @RequestMapping(value = "/tripCountFuelDateForm")
@@ -141,17 +152,70 @@ public class TripController {
     }
 
     @RequestMapping(value = "/tripCountFuelDriver")
-    public ModelAndView countFuelByDateView(@RequestParam("DriverName") String driverName) {
-        LOGGER.debug("get sumFuel for driver " + driverName);
-        List<Trip> trips = tripService.getTripsByDriver(driverName);
-        Double sum = 0d;
-        for (Trip current : trips) {
-            sum += current.getSumFuel();
+    public ModelAndView countFuelByDateView(@RequestParam("DriverId") String driverId) {
+        LOGGER.debug("get sumFuel for driver with ID = " + driverId);
+        Driver driver = driverService.getDriverById(Long.parseLong(driverId));
+        if (driver != null) {
+            List<Trip> trips = tripService.getTripsByDriver(driver.getName());
+            Double sum = 0d;
+            for (Trip current : trips) {
+                sum += current.getSumFuel();
+            }
+            Map<String, String> result = new HashMap<>();
+            result.put("driver", driver.getName());
+            Trip trip = new Trip();
+            result.put("sum", trip.getSumFuelString(sum));
+            return new ModelAndView("summaryByDriver", "result", result);
+        } else return getErrorPage("Водитель с ID = " + driverId + " не существует");
+    }
+
+    private Trip checkTripFields(Trip trip, String driverName, String startPoint, String endPoint, String distance,
+                                     String startDate, String endDate) {
+        boolean errorStatus = false;
+        Pattern isNumber = Pattern.compile("[0-9]+");
+
+        Driver driver = driverService.getDriverByName(driverName);
+        if (!errorStatus && driver != null)
+            trip.setDriver(driver);
+        else errorStatus = true;
+
+        if (!errorStatus && !isNumber.matcher(startPoint).matches())
+            trip.setStartPoint(startPoint);
+        else errorStatus = true;
+
+        if (!errorStatus && !isNumber.matcher(endPoint).matches())
+            trip.setEndPoint(endPoint);
+        else errorStatus = true;
+
+        try {
+            if (!errorStatus)
+                trip.setDistance(Double.parseDouble(distance));
+        } catch (NumberFormatException e) {
+            errorStatus = true;
         }
-        Map<String, String> result = new HashMap<>();
-        result.put("driver", driverName);
-        Trip trip = new Trip();
-        result.put("sum", trip.getSumFuelString(sum));
-        return new ModelAndView("summaryByDriver", "result", result);
+
+        Date start = null;
+        Date end = null;
+        try {
+            start = format.parse(startDate);
+            end = format.parse(endDate);
+            if (start.after(end)) errorStatus = true;
+        } catch (ParseException e) {
+            errorStatus = true;
+        }
+        if (!errorStatus) {
+            trip.setStartDate(start);
+            trip.setEndDate(end);
+            trip.setSumFuel();
+            return trip;
+        } else return null;
+    }
+
+    private ModelAndView getErrorPage(String exceptionMessage) {
+        Map<String, String> errorMap = new HashMap<>();
+        errorMap.put("datetime", format.format(new Date().getTime()));
+        errorMap.put("exception", exceptionMessage);
+        errorMap.put("url", "/drivers/driverInputForm");
+        return new ModelAndView("error/404", "errorMap", errorMap);
     }
 }
